requirements/development.txt
# Local development dependencies go here
-r base.txt

factory_boy


django-braces==1.10.0

django-debug-toolbar==1.6
coverage==4.2
codecov

